Deepen three shallow modules: PartyFinder composable, DifficultyRegistry, GuideIndex
- Extract usePartyFinder composable from PartyFinder.vue (643→20 line script)
  All API fetching, state, filtering logic, and types move to
  .vitepress/theme/hooks/usePartyFinder.ts; component becomes a thin renderer

- Replace difficultyTypes data array with DifficultyRegistry class
  Exposes find(), sortGuides(), sortedByHomeNav, all — callers no longer
  need to know the internal shape; adding a difficulty type now touches one file

- Extract guide-index.ts from navigation.ts (261→90 lines)
  discoverGuides() owns fs walking and frontmatter parsing; extractFrontmatter
  is private; navigation.ts becomes a slim VitePress orchestrator
